+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Change: Tweak descriptions to clarify that the links are to each author's post archive
 * Change: Add default values for optional arguments to inline parameter documentation
 * Change: Note compatibility through WP 6.6+
+* Change: Prevent unwarranted PHPCS complaints about unescaped output (HTML is allowed)
 * Change: Update copyright date (2024)
 * Change: Reduce number of 'Tags' from `readme.txt`
 * Change: Remove development and testing-related files from release packaging